requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
docx==0.2.4
fitz==0.0.1.dev2
Flask==3.0.3
flask_bcrypt==1.0.1
flask_login==0.6.3
flask_migrate==4.0.7
flask_sqlalchemy==3.1.1
MarkupSafe==2.1.3
Pillow
PyPDF2==3.0.1
Requests==2.32.3
scikit_learn==1.5.1
spacy==3.7.5
SQLAlchemy==2.0.32
Werkzeug==3.0.3
